+++ src/Resources/contao/languages/de/tl_fideid.php
@@ -45,6 +45,7 @@ $GLOBALS['TL_LANG']['tl_fideid']['elterneinverstaendnis'] = array('Zustimmung Er
 
 $GLOBALS['TL_LANG']['tl_fideid']['turnier_legend'] = 'Turnier';
 $GLOBALS['TL_LANG']['tl_fideid']['turnier'] = array('Turnier', 'Turnier, für das der Antrag gestellt wird.');
+$GLOBALS['TL_LANG']['tl_fideid']['turnierlink'] = array('', 'Link zur FIDE-Seite mit den registrierten Turnieren');
 
 $GLOBALS['TL_LANG']['tl_fideid']['germany_legend'] = 'Länderkennung GER';
 $GLOBALS['TL_LANG']['tl_fideid']['germany'] = array('GER', 'Länderkennung GER für den Antragsteller.');
@@ -56,6 +57,7 @@ $GLOBALS['TL_LANG']['tl_fideid']['intern_legend'] = 'Interne Bemerkungen';
 $GLOBALS['TL_LANG']['tl_fideid']['intern'] = array('Bemerkungen', 'Interne Bemerkungen');
 
 $GLOBALS['TL_LANG']['tl_fideid']['speedmail_legend'] = 'Schnelle E-Mail-Kommunikation';
+$GLOBALS['TL_LANG']['tl_fideid']['speedmail'] = array('', 'Buttons, mit denen Sie sofort eine E-Mail versenden können.');
 
 /**
  * Buttons für globale Operationen